✨ hide entity labels in thumbnail
